SAGE-1358: change overlay support to use GitHub cache
The CI and Release workflow are now responsible for executing the
overlay.sh script and caching the results into a GitHub cache. The
key for that cache is then passed to the reusable wildnode-image
workflows where the resulting overlayed file system is unpacked and used.

Note: this requires that the customize workflows run within the same
GitHub runner space (i.e. self-hosted) as the wildnode-image workflows.

reference: https://docs.github.com/en/actions/using-workflows/caching-dependencies-to-speed-up-workflows