SAGE-1358: add customized repo version information as a version extension on base image
The CI and Release workflows now compute the customized repo's version
information (from the `project_name` file and the `git tag` history) and
pass that as a "version extension" to the wildnode-image CI / Release
workflows. This results in the produced image files containing a
combination of the base image version and the customized version
information for proper tracking.